MLE-24405 Refactor: Moved some OkHttp-specific classes
Trying to get as many okhttp3 imports into one package as possible, with the notable exception for now of OkHttpServices.
