+++ marklogic-client-api/src/main/java/com/marklogic/client/impl/FailedRequest.java
@@ -11,7 +11,6 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import com.marklogic.client.io.Format;
-import okhttp3.MediaType;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
